Allow check for field length instead of contents
For dual-funded tests, we need a way to verify that witnesses are
the correct length instead of the correct contents. This is a bit
'complicated' because witnesses are in the range of 71-73 bytes long.

This patch allows a byte length (or range) to be specified by using
`*73` or `*71-73`, as in the case for a range of valid lengths.